# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "wlanpi-core",
...
    "version": "2.1.10"
  },
  "paths": {
...
    "/api/v1/wifi/capabilities": {
      "get": {
        "tags": [
          "wifi"
        ],
        "summary": "Show Wifi Capabilities",
        "description": "Return ``iw phy`` capability dumps for each PHY.",
        "operationId": "show_wifi_capabilities_api_v1_wifi_capabilities_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WifiCapabilitiesResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/wifi/regulatory": {
      "get": {
        "tags": [
          "wifi"
        ],
        "summary": "Show Wifi Regulatory",
        "description": "Return WiFi regulatory domain information.",
        "operationId": "show_wifi_regulatory_api_v1_wifi_regulatory_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WifiRegulatoryResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/wifi/hotspot/stations": {
      "get": {
        "tags": [
          "wifi"
        ],
        "summary": "Hotspot AP station list",
        "description": "Connected stations on the hotspot AP interface.\n\nReturns 409 when the device is not in hotspot mode.",
        "operationId": "show_hotspot_stations_api_v1_wifi_hotspot_stations_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "iface",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Iface"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HotspotStationsResponse"
                }
              }
            }
          },
          "409": {
            "description": "Device mode precondition not met (e.g. not in hotspot mode)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/wifi/hotspot/link": {
      "get": {
        "tags": [
          "wifi"
        ],
        "summary": "Hotspot per-client link stats",
        "description": "Per-station link statistics for hotspot AP clients.\n\nReturns 409 when the device is not in hotspot mode.",
        "operationId": "show_hotspot_client_link_api_v1_wifi_hotspot_link_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "iface",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Iface"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HotspotClientLinkResponse"
                }
              }
            }
          },
          "409": {
            "description": "Device mode precondition not met (e.g. not in hotspot mode)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HotspotClientLink": {
        "properties": {
          "mac": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mac"
          },
          "interface": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interface"
          },
          "signal_dbm": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signal Dbm"
          },
          "tx_bitrate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tx Bitrate"
          },
          "rx_bitrate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rx Bitrate"
          },
          "inactive_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inactive Time"
          },
          "connected_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Connected Time"
          },
          "authorized": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authorized"
          },
          "authenticated": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authenticated"
          },
          "associated": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Associated"
          }
        },
        "type": "object",
        "title": "HotspotClientLink"
      },
      "HotspotClientLinkResponse": {
        "properties": {
          "mode": {
            "type": "string",
            "title": "Mode",
            "example": "hotspot"
          },
          "interface": {
            "type": "string",
            "title": "Interface",
            "example": "wlan0"
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "example": 1
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/HotspotClientLink"
            },
            "type": "array",
            "title": "Links"
          }
        },
        "type": "object",
        "required": [
          "mode",
          "interface",
          "count"
        ],
        "title": "HotspotClientLinkResponse"
      },
...
      "HotspotStation": {
        "properties": {
          "mac": {
            "type": "string",
            "title": "Mac"
          },
          "interface": {
            "type": "string",
            "title": "Interface"
          },
          "signal_dbm": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signal Dbm"
          },
          "tx_bitrate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tx Bitrate"
          },
          "rx_bitrate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rx Bitrate"
          },
          "inactive_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inactive Time"
          },
          "connected_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Connected Time"
          },
          "authorized": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authorized"
          },
          "authenticated": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authenticated"
          },
          "associated": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Associated"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "mac",
          "interface"
        ],
        "title": "HotspotStation"
      },
      "HotspotStationsResponse": {
        "properties": {
          "mode": {
            "type": "string",
            "title": "Mode",
            "example": "hotspot"
          },
          "interface": {
            "type": "string",
            "title": "Interface",
            "example": "wlan0"
          },
          "count": {
            "type": "integer",
            "title": "Count",
            "example": 1
          },
          "stations": {
            "items": {
              "$ref": "#/components/schemas/HotspotStation"
            },
            "type": "array",
            "title": "Stations"
          }
        },
        "type": "object",
        "required": [
          "mode",
          "interface",
          "count"
        ],
        "title": "HotspotStationsResponse"
      },
...
      "MessageResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Human-readable error or status message"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "MessageResponse"
      },
...
      "PhyCapabilities": {
        "properties": {
          "phy": {
            "type": "string",
            "title": "Phy",
            "example": "phy0"
          },
          "info": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Info",
            "description": "Raw iw phy info output"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "phy"
        ],
        "title": "PhyCapabilities"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WifiCapabilitiesResponse": {
        "properties": {
          "adapters": {
            "items": {
              "$ref": "#/components/schemas/PhyCapabilities"
            },
            "type": "array",
            "title": "Adapters"
          }
        },
        "type": "object",
        "title": "WifiCapabilitiesResponse"
      },
      "WifiRegulatoryResponse": {
        "properties": {
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country",
            "example": "GB"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source",
            "example": "iw"
          },
          "raw": {
            "type": "string",
            "title": "Raw",
            "description": "Raw output of iw reg get"
          }
        },
        "type": "object",
        "required": [
          "raw"
        ],
        "title": "WifiRegulatoryResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}